Ship prebuilt GN binaries
Google stopped supporting 32 bit builds, and the build on older Ubuntu
versions is rather brittle. So we go the lazy route and add some
prebuilt binaries. If those don't work, the build script will try to
build GN itself, which may or may not work, depending on your binary
toolchain.