Fix errors in turn credential generation (#46)
* Changed parameter to be private

The parameters for the ICE Server Provider are not likely to be needed
by anything else. There is no need to be cluttering up the global
parameter space.

* Added uri into turn definition.

The URI was not being set on turn server definitions

* Significant fixes to REST call for getting turn credentials

The REST call was not looking for data in a valid field and was
missing appropriate checks for failed requests. Also added in
logging to allow future diagnostics